Enable file upload.
Uses fine-uploader (https://github.com/FineUploader/fine-uploader).
Minified JS lifted from fine-uploader's npm package.

Two requests are made to the server when the user hits submit. The first
uploads the file and the second initiates search on the uploaded file:

1. Tempfile object capturing the uploaded file is stored in a Hash,
   keyed by file name.
2. The name of the uploaded file is sent to the server and is used
   to populate params[:seq].

Hereafter things work like before - as if the user pasted the sequences.

The selected file is uploaded in one-piece, which is fine-uploader's
default behaviour. Concurrent chunked uploads is possible and can be
enabled later.

close #1

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>